+++ src/main/java/org/urbcomp/startdb/selfstar/compressor32/ElfStarCompressor32.java
@@ -2,6 +2,8 @@ package org.urbcomp.startdb.selfstar.compressor32;
 
 import org.urbcomp.startdb.selfstar.compressor32.xor.IXORCompressor32;
 import org.urbcomp.startdb.selfstar.utils.Elf32Utils;
+import org.urbcomp.startdb.selfstar.utils.Huffman.Code;
+import org.urbcomp.startdb.selfstar.utils.Huffman.HuffmanEncode;
 import org.urbcomp.startdb.selfstar.utils.OutputBitStream;
 
 import java.util.Arrays;
@@ -13,9 +15,11 @@ public class ElfStarCompressor32 implements ICompressor32 {
     private final int[] leadDistribution = new int[32];
     private final int[] trailDistribution = new int[32];
     private OutputBitStream os;
-    private long compressedSizeInBits = 0;
+    private int compressedSizeInBits = 0;
     private int lastBetaStar = Integer.MAX_VALUE;
     private int numberOfValues = 0;
+    private final int[] frequency = new int[9];    // 0 is for 10-i, 16 is for not erasing
+    private Code[] huffmanCode;
 
     public ElfStarCompressor32(IXORCompressor32 xorCompressor) {
         this.xorCompressor = xorCompressor;
@@ -30,9 +34,11 @@ public class ElfStarCompressor32 implements ICompressor32 {
         if (v == 0.0 || Float.isInfinite(v)) {
             vPrimeList[numberOfValues] = vInt;
             betaStarList[numberOfValues] = Integer.MAX_VALUE;
+            frequency[8]++;
         } else if (Float.isNaN(v)) {
             vPrimeList[numberOfValues] = 0x7fc00000;
             betaStarList[numberOfValues] = Integer.MAX_VALUE;
+            frequency[8]++;
         } else {
             // C1: v is a normal or subnormal
             int[] alphaAndBetaStar = Elf32Utils.getAlphaAndBetaStar(v, lastBetaStar);
@@ -42,14 +48,14 @@ public class ElfStarCompressor32 implements ICompressor32 {
             int mask = 0xffffffff << eraseBits;
             int delta = (~mask) & vInt;
             if (delta != 0 && eraseBits > 3) {  // C2
-                if (alphaAndBetaStar[1] != lastBetaStar) {
-                    lastBetaStar = alphaAndBetaStar[1];
-                }
+                lastBetaStar = alphaAndBetaStar[1];
                 betaStarList[numberOfValues] = lastBetaStar;
                 vPrimeList[numberOfValues] = mask & vInt;
+                frequency[lastBetaStar]++;
             } else {
                 betaStarList[numberOfValues] = Integer.MAX_VALUE;
                 vPrimeList[numberOfValues] = vInt;
+                frequency[8]++;
             }
         }
         numberOfValues++;
@@ -68,16 +74,14 @@ public class ElfStarCompressor32 implements ICompressor32 {
     }
 
     private void compress() {
+        huffmanCode = HuffmanEncode.getHuffmanCodes(frequency);
+        compressedSizeInBits += HuffmanEncode.writeHuffmanCodes(os, huffmanCode);
         xorCompressor.setDistribution(leadDistribution, trailDistribution);
-        lastBetaStar = Integer.MAX_VALUE;
         for (int i = 0; i < numberOfValues; i++) {
             if (betaStarList[i] == Integer.MAX_VALUE) {
-                compressedSizeInBits += os.writeInt(2, 2); // case 10
-            } else if (betaStarList[i] == lastBetaStar) {
-                compressedSizeInBits += os.writeBit(false);    // case 0
+                compressedSizeInBits += os.writeInt(huffmanCode[8].code, huffmanCode[8].length); // not erase
             } else {
-                compressedSizeInBits += os.writeInt(betaStarList[i] | 0x18, 5);  // case 11, 2 + 3 = 5
-                lastBetaStar = betaStarList[i];
+                compressedSizeInBits += os.writeInt(huffmanCode[betaStarList[i]].code, huffmanCode[betaStarList[i]].length);  // case 11, 2 + 4 = 6
             }
             compressedSizeInBits += xorCompressor.addValue(vPrimeList[i]);
         }
@@ -101,12 +105,8 @@ public class ElfStarCompressor32 implements ICompressor32 {
         calculateDistribution();
         compress();
         // we write one more bit here, for marking an end of the stream.
-        compressedSizeInBits += os.writeInt(2, 2);  // case 10
+        compressedSizeInBits += os.writeInt(huffmanCode[8].code, huffmanCode[8].length); // not erase
         compressedSizeInBits += xorCompressor.close();
-    }
-
-    public String getKey() {
-        return xorCompressor.getKey();
     }
 
     public void refresh() {
@@ -115,6 +115,7 @@ public class ElfStarCompressor32 implements ICompressor32 {
         lastBetaStar = Integer.MAX_VALUE;
         numberOfValues = 0;
         os = xorCompressor.getOutputStream();
+        Arrays.fill(frequency, 0);
         Arrays.fill(leadDistribution, 0);
         Arrays.fill(trailDistribution, 0);
     }
