+++ src/main/java/org/urbcomp/startdb/selfstar/decompressor/ElfStarDecompressor.java
@@ -1,10 +1,95 @@
 package org.urbcomp.startdb.selfstar.decompressor;
 
 import org.urbcomp.startdb.selfstar.decompressor.xor.IXORDecompressor;
+import org.urbcomp.startdb.selfstar.utils.Elf64Utils;
+import org.urbcomp.startdb.selfstar.utils.Huffman.Code;
+import org.urbcomp.startdb.selfstar.utils.Huffman.HuffmanEncode;
+import org.urbcomp.startdb.selfstar.utils.Huffman.Node;
+import org.urbcomp.startdb.selfstar.utils.InputBitStream;
 
-// all ElfStar (batch and stream) share the same decompressor
-public class ElfStarDecompressor extends org.urbcomp.startdb.selfstar.decompressor.ElfPlusDecompressor {
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+
+public class ElfStarDecompressor implements IDecompressor {
+    private static Code[] huffmanCode = new Code[17];
+    private final IXORDecompressor xorDecompressor;
+    private int lastBetaStar = Integer.MAX_VALUE;
+    private Node root;
+
     public ElfStarDecompressor(IXORDecompressor xorDecompressor) {
-        super(xorDecompressor);
+        this.xorDecompressor = xorDecompressor;
+    }
+
+    public List<Double> decompress() {
+        List<Double> values = new ArrayList<>(1024);
+        initHuffmanTree();
+        Double value;
+        while ((value = nextValue()) != null) {
+            values.add(value);
+        }
+        return values;
+    }
+
+    private void initHuffmanTree() {
+        HuffmanEncode.readHuffmanCodes(xorDecompressor.getInputStream(), huffmanCode);
+        root = HuffmanEncode.buildHuffmanTree(huffmanCode);
+    }
+
+    @Override
+    public void refresh() {
+        lastBetaStar = Integer.MAX_VALUE;
+        xorDecompressor.refresh();
+        huffmanCode = new Code[17];
+    }
+
+    @Override
+    public void setBytes(byte[] bs) {
+        this.xorDecompressor.setBytes(bs);
+    }
+
+    @Override
+    public Double nextValue() {
+        Double v;
+        Node current = root;
+        while (true) {
+            current = current.children[readInt(1)];
+            if (current.data >= 0) {
+                if (current.data != 16) {
+                    lastBetaStar = current.data;
+                    v = recoverVByBetaStar();
+                } else {
+                    v = xorDecompressor.readValue();
+                }
+                break;
+            }
+        }
+        return v;
+    }
+
+    private Double recoverVByBetaStar() {
+        double v;
+        Double vPrime = xorDecompressor.readValue();
+        int sp = Elf64Utils.getSP(Math.abs(vPrime));
+        if (lastBetaStar == 0) {
+            v = Elf64Utils.get10iN(-sp - 1);
+            if (vPrime < 0) {
+                v = -v;
+            }
+        } else {
+            int alpha = lastBetaStar - sp - 1;
+            v = Elf64Utils.roundUp(vPrime, alpha);
+        }
+        return v;
+    }
+
+    @SuppressWarnings("all")
+    private int readInt(int len) {
+        InputBitStream in = xorDecompressor.getInputStream();
+        try {
+            return in.readInt(len);
+        } catch (IOException e) {
+            throw new RuntimeException("IO error: " + e.getMessage());
+        }
     }
 }
